service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
